Extract math and slices utility packages
Move the Min function to a new math package

Move the Map and Sum function to a new slices package

We import these with the names slcs and mth so they don't clash with the
standard libraries.   Would prefer if the slices package could be made
to just import/re-export the standard package so we're not having to
deal with multiple names, but that's not how golang do.