Remove some []byte() shenanigans
Felt like a better idea as we wrote is, as range string iterates over
(int, rune) and indexing/slicing a string returns bytes.

As we're using all ASCII strings in our puzzle inputs, this works out
that

for i, r := range input

Will make i go 0,1,2,3,4,5,6,7,8...

But in the case of unicode input it might actually go 0, 3, 4

To avoid this complexity/overhead we might be better off loading inputs
into []byte slices instead.