Day 15 Parts 1 and 2
There's something quite unsatisfactory here.

In order to generate the space of values we've got two hardcoded
functions, one that solves for two ingredients, and one that solves for
four.  This would be better served by a function to generate the
partitions(100, len(ingredients)) but I couldn't come up with a clear
version.