Day 19 Part 2: Greedily decompose the result
Part 2 is to run the 'machine' backwards, so we need to map from
expansion, back to replacement.

If we apply this greedily by replacing the longest expansion that
matches each step then this solves for our input data.  This isn't very
general purpose, for that we'd want to backtrack maybe and need more
state, but I think just solving for the input at hand is valid.