2017:1:2: uncaptcha_half
Second verse, same as the first, but with the stride for comparison
being len(puzzle)/2 rather than 1.

Could generalise this to a function that takes a stride function, Part1
is then i -> i+1, and 2 is i > i+len(puzzle)/2, but that's not the best
tradeoff in terms of complexity/loc.