2017:3:2 Calculate the neighbour sums
We split the existing spiralManattan function to take a predicate to
indicate if it should keep search.  The predicate takes the x,y
coordinates and returns a bool.  For the original case we can ignore the
arguments and just track that we've been called n times.

For part 2 we can supply a more interesting predicate that computes the
neighbour sums in a closed-over hash, and will stop looking once the sum
exceeds the puzzle input.