Change testing helper from deep to cmp
I wanted to check that the test for buildTree was good so I broke the
data and the test didn't fail?!

Suspecting deep, I swearched around for an alternative and hit on
https://github.com/google/go-cmp so tweaked the test to use it instead.

It gave the illuminating error message:

    panic: cannot handle unexported field at {day07.Node}.name:
            "github.com/richardc/advent-go/2017/day07".Node
    consider using a custom Comparer; if you control the implementation of type, you can also consider using an Exporter, AllowUnexported, or cmpopts.IgnoreUnexported [recovered]

So the fix is to make the fields public.  Doing that makes deep.Equal
reflect the failing test too.

Going back and forwards, I found the diff output from cmp to be more
pleasing, so am going to stay switched over to that.