2017:13:2 Figure out the safeTimeToGo
Changes over to a computed state for the firewall rather than a
similuated one.  It's very like the sphere drop machine from the
previous year, but the period of the scanner reaching 0 is 2*(max-1)

The safeTimeToGo is when Firewall.Triggered(time) == false.

This is different from a firewall.Severity == 0 as you could trigger the
0th column, leading to a severity of 0 but still a tripped firewall. #
Please enter the commit message for your changes. Lines starting