2017:14:2 Count the contiguous islands
After finishing part 1 I'd guessed part 2 would be pathfinding, but
instead it's pool/region finding.

We solved a very similar problem for 2021/day9/part2, so this is close
to a translation of that.  We do a linear scan of the map, and then if we
find a true value we send out a visitor on a recursive walk that mutates
the map and turns off all the connected neighbours.

https://github.com/richardc/advent-rust-2021/blob/main/src/day09.rs#L54