2017:16:2 Dance 1 billion times
Well since we may well permute through the start state, we only need to
compute the few dances that lead back to that state, then the remainder
after fast forwarding the search.

That's 59 iterations for our sample data.  We do technically
overcompute, as we'll calculate the tail again.

Some memoisation would trade memory for cpu, bt it's not that much to
save here.