2017:17:2 Simulate just 0 moving
For part 2 we're allowed to just consider what's after 0;  after 50
million iterations.   Time to crack out the maths.

Adding InspectValueAfter0After to create a table, we can see the numbers
jump up with a fun pattern.

1 1
2 1
3 3
...
9 3
10 10
...
44 10
45 45
46 45
47 45
48 45
49 45
50 50

Looking at that and taking a bit of extra thinking we can see it's going
up when the insertion point would be == 0.  We can simulate that with a
simple for loop, and that's acceptably quick (~2.3seconds) compared to
deriving the true function.