2017:18:2 Duet
To make our CPU duet we change sleep to be a string, the name of the
register to take the data into we are woken.  Then we have a TryWake
which, if we are asleep, and now have something in our input channel,
populates the register, updates the channel, and clears the sleep, so
next loop round we can Run() again.