2017:23:2 Optimise away the body
On inspecting the assembly we see it's a very ineffient prime filter,
checking every 17th value between 106700 and 123700 and recording the
composite values.

Assuming those are the numbers unique to our puzzle, we extract the
17 from the program source, and then run the first 8 instructions
to extract the bounds from registers b and c.

Then it's a simple case of using math/big's ProbablyPrime in a loop over
the extracted range.

Feels a bit cheaty, but it's far easier than adding a peephole optimiser
to swap in a mod operation for the sub loop.