2017:24:1 Chain building graph walker
Another fun graph traversal.

Mild fight with go again,

  func (c Chain) Append(s Section) Chain {
	  return Chain{
		  Links: append(c.Links, s),
	  }
  }

Seems like it would be sufficient to make a clone with the extra item
appended, but the c.Links slice is by reference, so we sometimes end up
with the Links overwriting each others tails.  A full copy is needed, to
avoid the accidental mutation, so:

    append([]Section(nil), append(c.Links, s)...)

Runtime is a little high, probably memory pressure, as we keep all the
chains and evaluate them at the end.  We'll see if part 2 makes this an
issue.