2019:6:2 Calculate transfers between YOU and SAN
We calculate the steps for all the nodes up to the root for both
ourselves and santa.  For common nodes the path is the two paths added
up.  Find the minimal value.

There's probably some optimisations to be had here in returning the
paths from one of the paricipants in order so we can stop early, but
full scans are pretty cheap.