2019:7:2 Block on empty input
Here we make two small changed to the intcode intepreter - first it
takes an optional function to handle output.  We use these to link 5
interpreters together.

The other key change is to make the input operator shift into a blocked
state if there is no input ready to be consumed.  We clear this blocked
flag in the input function.

Then after constructing our chained and looping amplifiers, we run them
until all cpus become halted.