2019:9:1 Add relative addressing to intcode and self-test
Nice puzzle in as much as it outputs actionable diagnostics if you've
mis-implemented the addressing modes.

intcode.Cpu#Address added as with the introdution of relative adressing,
we also need to allow the write instructions use it - previous
specification had the write argument never be immediate, so we didn't do
the lookup for it.